+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased - will be 1.2.0]
 ### Added
 - Full compatibility with [APCu Manager](https://wordpress.org/plugins/apcu-manager/).
+- Full integration with PerfOps.One suite.
 ### Changed
+- New menus (in the left admin bar) for accessing features: "PerfOps Analytics", "PerfOps Tools" and "PerfOps Settings".
 - In lists, it's now possible to navigate by direct page input.
 - Analysis delta time has been increased to avoid holes in stats when cron is not fully reliable.
 ### Fixed